requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.0
Flask-JWT-Extended==4.4.0
Flask-CORS==4.0.0
PyMySQL==1.0.2
Werkzeug==2.3.0
python-dotenv==1.0.0
APScheduler==3.10.0
